# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "角色扮演对话API",
    "description": "提供角色扮演和聊天功能的后端 API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "注册新用户",
        "description": "使用邮箱、密码与可选的用户名注册一个新账号。若邮箱已被注册则返回 400。\n\n成功后返回新建用户的公开信息（不含密码哈希）。",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "新创建的用户公开信息。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "400": {
            "description": "邮箱已被注册或请求体校验失败。"
          },
          "422": {
            "description": "请求体字段类型或格式不合法（如邮箱格式错误、密码为空）。"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "用户登录并获取 JWT",
        "description": "使用邮箱与密码换取 JWT access token。\n\n- 邮箱不存在或密码错误：返回 401。\n- 账号被禁用：返回 400。\n成功后返回的 token 默认有效期由后端配置（ACCESS_TOKEN_EXPIRE_MINUTES）控制。",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "JWT 访问令牌及其类型。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "账号已被禁用。"
          },
          "401": {
            "description": "邮箱不存在或密码错误。"
          },
          "422": {
            "description": "请求体字段类型或格式不合法。"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "获取当前登录用户信息",
        "description": "解析请求头中的 JWT 并返回当前登录用户的公开信息。",
        "operationId": "read_me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "当前登录用户公开信息。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "未提供有效 JWT，或 token 已过期/被吊销。"
          },
          "403": {
            "description": "凭据有效但当前被禁止访问该资源。"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "获取可用角色列表",
        "description": "返回当前用户可见的角色列表（`RoleOut` 数组）。\n\n- 内置角色（isBuiltin=True）始终可见。\n- 自定义角色（isMine=True）仅在已登录且属于当前用户时返回。\n- 未登录访问时，仅返回内置角色。",
        "operationId": "get_roles_api_roles_get",
        "responses": {
          "200": {
            "description": "RoleOut 数组，字段为 camelCase。",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleOut"
                  },
                  "type": "array",
                  "title": "Response Get Roles Api Roles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/roles/my": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "创建当前用户的自定义角色",
        "description": "为当前登录用户创建一个新的自定义角色。\n\n- `name` 在同一用户下唯一（重复返回 400）。\n- `default_speaker_id` 可选；未设置时后续 TTS 调用须显式传入 speakerId。\n- 头像请通过 `POST /roles/{roleId}/avatar` 单独上传。",
        "operationId": "create_my_role_api_roles_my_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "新创建的自定义角色公开信息。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "400": {
            "description": "角色名重复，或请求体字段类型不合法。"
          },
          "401": {
            "description": "未提供有效 JWT。"
          },
          "422": {
            "description": "请求体字段类型或长度不合法。"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/roles/my/{role_id}": {
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "更新当前用户的自定义角色",
        "description": "局部更新当前用户的指定自定义角色。仅修改提供的字段；未提供的字段保持不变。\n\n- 不允许修改内置角色。\n- 若 roleId 不存在或不属于当前用户：返回 404。",
        "operationId": "update_my_role_api_roles_my__role_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Role Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的角色公开信息。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "404": {
            "description": "角色不存在或当前用户无权修改。"
          },
          "422": {
            "description": "请求体字段类型不合法。"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "删除当前用户的自定义角色",
        "description": "硬删除当前用户拥有的指定自定义角色。不允许删除内置角色。\n\n- 若 roleId 不存在或不属于当前用户：返回 404。",
        "operationId": "delete_my_role_api_roles_my__role_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功。"
          },
          "404": {
            "description": "角色不存在或当前用户无权删除。"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/roles/{role_id}/avatar": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "上传并更新角色头像",
        "description": "上传一张图片作为指定自定义角色的头像，支持 image/* MIME（PNG/JPEG/WebP）。\n\n- 仅允许修改当前用户拥有的自定义角色。\n- 若文件不是图片：返回 400。\n- 头像二进制存储于数据库（avatar_blob），获取请用 `GET /roles/{roleId}/avatar`。",
        "operationId": "upload_role_avatar_api_roles__role_id__avatar_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Role Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_role_avatar_api_roles__role_id__avatar_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新头像后的角色公开信息（含 avatar_url）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "400": {
            "description": "上传文件不是图片，或请求体不合法。"
          },
          "404": {
            "description": "角色不存在或当前用户无权修改。"
          },
          "413": {
            "description": "文件过大（超出后端限制）。"
          },
          "422": {
            "description": "表单字段类型不合法。"
          }
        }
      },
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "获取角色头像二进制",
        "description": "直接返回指定角色的头像图片原始字节，Content-Type 与上传时一致（如 image/png）。\n\n- 若角色不存在或未上传头像：返回 404。\n- 该端点不需要鉴权（头像可公开）。",
        "operationId": "get_role_avatar_api_roles__role_id__avatar_get",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "头像二进制（image/*）。",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "角色不存在或尚未上传头像。"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_role_avatar_api_roles__role_id__avatar_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_role_avatar_api_roles__role_id__avatar_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "description": "角色名（同一用户下唯一）。",
            "examples": [
              "百战堂"
            ]
          },
          "systemPrompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Systemprompt",
            "description": "LLM 系统提示词，可选。",
            "examples": [
              "你是一位身经百战的老将。"
            ]
          },
          "defaultSpeakerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultspeakerid",
            "description": "TTS 默认 speaker ID，可选。",
            "examples": [
              "spk_male_01"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RoleCreate",
        "description": "创建自定义角色请求体。"
      },
      "RoleOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "角色主键 ID。",
            "examples": [
              1
            ]
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug",
            "description": "URL 友好 slug，内置角色固定；自定义角色由后端生成。",
            "examples": [
              "baizhantang"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "展示用角色名。",
            "examples": [
              "百战堂"
            ]
          },
          "systemPrompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Systemprompt",
            "description": "注入 LLM 的系统提示词；为空时使用后端默认。",
            "examples": [
              "你是一位身经百战的老将，言语简洁。"
            ]
          },
          "defaultSpeakerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultspeakerid",
            "description": "TTS 默认 speaker ID；未设置则 TTS 调用须显式传入 speakerId。",
            "examples": [
              "spk_male_01"
            ]
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl",
            "description": "头像 URL（指向 /roles/{id}/avatar），未上传时为 null。",
            "examples": [
              "/api/roles/1/avatar"
            ]
          },
          "isBuiltin": {
            "type": "boolean",
            "title": "Isbuiltin",
            "description": "是否系统内置角色（内置不可修改/删除）。",
            "examples": [
              true
            ]
          },
          "isMine": {
            "type": "boolean",
            "title": "Ismine",
            "description": "是否当前用户创建的自定义角色。",
            "default": false,
            "examples": [
              false
            ]
          },
          "inGame": {
            "type": "boolean",
            "title": "Ingame",
            "description": "是否可作为世界实体进入游戏。",
            "default": false,
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "isBuiltin"
        ],
        "title": "RoleOut",
        "description": "角色对外公开信息。"
      },
      "RoleUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "新角色名。",
            "examples": [
              "百战堂（重命名）"
            ]
          },
          "systemPrompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Systemprompt",
            "description": "新的系统提示词。",
            "examples": [
              "你是一位身经百战的老将，言语中带江湖气。"
            ]
          },
          "defaultSpeakerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultspeakerid",
            "description": "新的默认 speaker ID。",
            "examples": [
              "spk_male_02"
            ]
          }
        },
        "type": "object",
        "title": "RoleUpdate",
        "description": "更新自定义角色请求体（局部更新）。"
      },
...
      "Token": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken",
            "description": "JWT access token，前端放入 Authorization: Bearer <token> 头。",
            "examples": [
              "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            ]
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "description": "Token 类型，固定为 \"bearer\"。",
            "default": "bearer",
            "examples": [
              "bearer"
            ]
          }
        },
        "type": "object",
        "required": [
          "accessToken"
        ],
        "title": "Token",
        "description": "登录成功响应：JWT access token。"
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "用户邮箱，作为登录账号，必须为合法邮箱格式。",
            "examples": [
              "alice@example.com"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "明文密码（服务端使用 bcrypt 加盐哈希后存储）。",
            "examples": [
              "S3cret!Pass"
            ]
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "可选显示名；为空时使用邮箱前缀作为默认。",
            "examples": [
              "alice"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserCreate",
        "description": "注册新用户请求体。"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "注册时使用的邮箱。",
            "examples": [
              "alice@example.com"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "明文密码。",
            "examples": [
              "S3cret!Pass"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "description": "用户登录请求体。"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "用户主键 ID。",
            "examples": [
              1
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "用户邮箱。",
            "examples": [
              "alice@example.com"
            ]
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "显示名。",
            "examples": [
              "alice"
            ]
          },
          "isAdmin": {
            "type": "boolean",
            "title": "Isadmin",
            "description": "是否管理员账号；管理员拥有后台管理权限。",
            "examples": [
              false
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "isAdmin"
        ],
        "title": "UserOut",
        "description": "用户公开信息（不含密码哈希）。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  },
...
}